agrisense_app/backend/requirements.safe.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
paho-mqtt==2.1.0
python-dotenv==1.0.1
requests
lightgbm
pyserial>=3.5
PyYAML==6.0.2
tensorflow==2.16.1
Flask==3.0.3
pymongo==4.8.0
google-generativeai
openai
rank-bm25==0.2.2

# PyTorch and sentence-transformers for direct SBERT usage
torch>=2.0.0
sentence-transformers>=2.2.0

# Enhanced backend dependencies (safe subset)
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.35
redis[hiredis]==5.2.1
fastapi-users[sqlalchemy]==14.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
alembic==1.14.0

# Celery and background task processing (leave out redis-py-cluster)
celery[redis]==5.4.0
flower==2.0.1
kombu==5.4.2

# Additional monitoring and task management
prometheus-client==0.21.1
psutil==6.1.0
